moved town demo into separate bin
