+++ src/bin/town_demo.rs
@@ -1,7 +1,5 @@
 extern crate mercurylib;
-
-use mercurylib::game::Game;
-use mercurylib::platform::Platform;
+extern crate image as im;
 
 #[cfg(target_family = "unix")]
 use mercurylib::desktopplatform::DesktopPlatform as CurrentPlatform;
@@ -9,9 +7,84 @@ use mercurylib::desktopplatform::DesktopPlatform as CurrentPlatform;
 //#[cfg(target_family = "wasm")]
 //use mercurylib::wasmplatform::WasmPlatform as CurrentPlatform;
 
+use mercurylib::platform::Platform;
+use mercurylib::background::Background;
+use mercurylib::player::Player;
+//use crate::input::Input;
+use mercurylib::gameobject::GameObject;
+use mercurylib::collision::Rect;
+use mercurylib::{SCREEN_WIDTH, SCREEN_HEIGHT};
+use std::time::Instant;
+
+pub struct TownDemoGame {
+    //pub input: Input,
+    gameobjects: Vec<Box<dyn GameObject>>
+}
+
+impl TownDemoGame {
+    pub fn new() -> Self {
+        let player = Box::new(Player::new());
+
+        //let mytilemap = Box::new(TileMap::new(2));
+        let bg = Box::new(Background::new("map.jpg"));
+
+        // setup subscriptions
+        //let mut input = Input::new();
+        //input.subscribe(player.observer.clone(), vec!["up", "down", "left", "right"]);
+        //input.subscribe(bg.observer.clone(), vec!["up", "down", "left", "right"]);
+
+        TownDemoGame {
+            //input: input,
+            gameobjects: vec![bg, player]
+        }
+    }
+}
+
+impl GameObject for TownDemoGame {
+    // create the screen image
+    fn render(&self) -> Option<im::RgbaImage> {
+        let mut screen_img = im::RgbaImage::new(SCREEN_WIDTH as u32, SCREEN_HEIGHT as u32);
+
+        for g in &self.gameobjects {
+            
+            if let Some(img) = g.render() {
+                if let Some(pos) = g.position() {
+                    let start = Instant::now();
+
+                    #[cfg(target_family = "unix")]
+                    im::imageops::overlay(&mut screen_img, &img, pos.x as i64, pos.y as i64);
+                    #[cfg(target_family = "wasm")]
+                    im::imageops::overlay(&mut screen_img, &img, pos.x as u32, pos.y as u32);
+                    
+                    let duration = start.elapsed();
+                    println!("overlay: {:?}", duration);
+                } else {
+                    continue;
+                }
+            } else {
+                continue;
+            }
+            
+        }
+
+        return Some(screen_img);
+    }
+
+    fn position(&self) -> Option<Rect> {
+        return Some(Rect{x: 0.0, y: 0.0, w: SCREEN_WIDTH as f64, h: SCREEN_HEIGHT as f64});
+    }
+
+    fn update(&mut self) {
+        for u in self.gameobjects.iter_mut() {
+            u.update();
+        }
+    }
+}
+
+
 #[cfg(target_family = "unix")]
 fn main() {
-    let mut game = Game::new();
+    let mut game = TownDemoGame::new();
     let mut pf = CurrentPlatform::new();
     pf.gameloop(&mut game);
 }
