+++ src/platform.rs
@@ -1,8 +1,8 @@
-use crate::game::Game;
+use crate::gameobject::GameObject;
 
 /*
  * trait to hide platform implementation behind
  */
 pub trait Platform {
-    fn gameloop(&mut self, game: &mut Game);
+    fn gameloop(&mut self, game: &mut dyn GameObject);
 }
