+++ src/player.rs
@@ -3,7 +3,7 @@ use crate::sprite::Direction;
 use crate::observer::Listener;
 use crate::gameobject::GameObject;
 use crate::collision::Rect;
-use crate::game::{SCREEN_WIDTH, SCREEN_HEIGHT};
+use crate::{SCREEN_WIDTH, SCREEN_HEIGHT};
 
 use std::rc::Rc;
